feat: moved Counter Schema to helper file
